requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket support
websockets==12.0

# CORS middleware (already included in FastAPI, but explicit version)
# fastapi already includes starlette which has CORS support

# LlamaIndex and LlamaCloud services
llama-index==0.9.13
llama-cloud==0.0.6
llama-parse==0.4.4

# Alternative if the above don't work, try:
# llama-index-core
# llama-index-llms-openai
# llama-index-embeddings-openai

# Environment variables
python-dotenv==1.0.0

# File handling
python-multipart==0.0.6

# Additional dependencies that might be needed
pydantic==2.5.0
typing-extensions==4.8.0

# Optional: For better async support
asyncio-mqtt==0.13.0

# Optional: For logging
structlog==23.2.0